Get rid of cruft in implementation of execute_in_process. Fix structure of graph.execute_in_process
