+++ python_modules/dagster/dagster/core/definitions/graph.py
@@ -21,7 +21,11 @@ from dagster.core.definitions.config import ConfigMapping
 from dagster.core.definitions.definition_config_schema import IDefinitionConfigSchema
 from dagster.core.definitions.mode import ModeDefinition
 from dagster.core.definitions.resource import ResourceDefinition
-from dagster.core.errors import DagsterInvalidConfigError, DagsterInvalidDefinitionError
+from dagster.core.errors import (
+    DagsterInvalidConfigError,
+    DagsterInvalidDefinitionError,
+    DagsterInvariantViolationError,
+)
 from dagster.core.storage.io_manager import io_manager
 from dagster.core.types.dagster_type import (
     DagsterType,
@@ -55,7 +59,7 @@ if TYPE_CHECKING:
     from .partition import PartitionedConfig
     from .executor import ExecutorDefinition
     from .pipeline import PipelineDefinition
-    from dagster.core.execution.execute import InProcessGraphResult
+    from dagster.core.execution.execute_in_process import InProcessGraphResult
 
 
 def _check_node_defs_arg(graph_name: str, node_defs: List[NodeDefinition]):
@@ -507,7 +511,7 @@ class GraphDefinition(NodeDefinition):
 
     def execute_in_process(
         self,
-        run_config: Optional[Dict[str, Any]] = None,
+        config: Any = None,
         instance: Optional["DagsterInstance"] = None,
         resources: Optional[Dict[str, Any]] = None,
     ):
@@ -515,8 +519,8 @@ class GraphDefinition(NodeDefinition):
         Execute this graph in-process, collecting results in-memory.
 
         Args:
-            run_config (Optional[Dict[str, Any]]):
-                Configuration for the run.
+            config (Optional[Dict[str, Any]]):
+                Configuration for the graph.
             instance (Optional[DagsterInstance]):
                 The instance to execute against, an ephemeral one will be used if none provided.
             resources (Optional[Dict[str, Any]]):
@@ -526,15 +530,47 @@ class GraphDefinition(NodeDefinition):
         Returns:
             InProcessGraphResult
         """
-        from dagster.core.execution.execute import execute_in_process
+        from dagster.core.execution.execute_in_process import core_execute_in_process
+        from dagster.core.instance import DagsterInstance
+        from dagster.core.storage.io_manager import IOManagerDefinition, IOManager
+        from .pipeline import PipelineDefinition
+        from .executor import execute_in_process_executor
 
-        run_config = check.opt_dict_param(run_config, "run_config")
+        if len(self.input_defs) > 0:
+            raise DagsterInvariantViolationError(
+                "Graphs with inputs cannot be used with execute_in_process at this time."
+            )
 
-        return execute_in_process(
+        instance = check.opt_inst_param(instance, "instance", DagsterInstance)
+        resources = check.opt_dict_param(resources, "resources", key_type=str)
+
+        resource_defs = {}
+        # Wrap instantiated resource values in a resource definition.
+        # If an instantiated IO manager is provided, wrap it in an IO manager definition.
+        for resource_key, resource in resources.items():
+            if isinstance(resource, ResourceDefinition):
+                resource_defs[resource_key] = resource
+            elif isinstance(resource, IOManager):
+                resource_defs[resource_key] = IOManagerDefinition.hardcoded_io_manager(resource)
+            else:
+                resource_defs[resource_key] = ResourceDefinition.hardcoded_resource(resource)
+
+        in_proc_mode = ModeDefinition(
+            executor_defs=[execute_in_process_executor], resource_defs=resource_defs
+        )
+
+        ephemeral_pipeline = PipelineDefinition(
+            name=self._name, graph_def=self, mode_defs=[in_proc_mode]
+        )
+
+        run_config = {"ops": config if config is not None else {}}
+
+        return core_execute_in_process(
             node=self,
+            ephemeral_pipeline=ephemeral_pipeline,
             run_config=run_config,
             instance=instance,
-            resources=resources,
+            output_capturing_enabled=True,
         )
